Move the theme manipulation to theme context.
