+++ src-tauri/capabilities/default.json
@@ -1,4 +1,5 @@
 {
+  "$schema": "../gen/schemas/desktop-schema.json",
   "identifier": "main",
   "description": "The main permissions of the app",
   "local": true,
